[a bit cleanup]
