+++ src/lib.rs
@@ -15,7 +15,7 @@ use std::slice;
 use std::ptr::null_mut;
 use byte::BytesExt;
 use dash_spv_ffi::ffi::boxer::{boxed, boxed_vec};
-use dash_spv_ffi::ffi::callbacks::{AddInsightBlockingLookup, GetBlockHeightByHash, GetBlockHashByHeight, MasternodeListDestroy, MasternodeListLookup, ShouldProcessLLMQTypeCallback, ValidateLLMQCallback, MerkleRootLookup, MasternodeListSave, SaveLLMQSnapshot, GetLLMQSnapshotByBlockHash, LogMessage, HashDestroy, LLMQSnapshotDestroy, SendError, ShouldProcessDiffWithRange};
+use dash_spv_ffi::ffi::callbacks::{AddInsightBlockingLookup, GetBlockHeightByHash, GetBlockHashByHeight, MasternodeListDestroy, MasternodeListLookup, ShouldProcessLLMQTypeCallback, ValidateLLMQCallback, MerkleRootLookup, MasternodeListSave, SaveLLMQSnapshot, GetLLMQSnapshotByBlockHash, LogMessage, HashDestroy, LLMQSnapshotDestroy, ShouldProcessDiffWithRange};
 use dash_spv_ffi::ffi::to::ToFFI;
 use dash_spv_ffi::ffi::unboxer::{unbox_any, unbox_block, unbox_qr_info_result, unbox_llmq_snapshot, unbox_llmq_validation_data, unbox_mn_list_diff_result, unbox_masternode_list};
 use dash_spv_ffi::types;
@@ -23,7 +23,7 @@ use dash_spv_models::llmq;
 use dash_spv_models::masternode::LLMQEntry;
 use dash_spv_primitives::consensus::encode;
 use dash_spv_primitives::crypto::byte_util::BytesDecodable;
-use crate::processing::{MasternodeProcessor, MNListDiffResult, MasternodeProcessorCache, QRInfoResult, ProcessingError};
+use crate::processing::{MasternodeProcessor, MasternodeProcessorCache, ProcessingError};
 
 /// Destroys anonymous internal holder for UInt256
 #[no_mangle]
@@ -68,9 +68,6 @@ pub unsafe extern fn processor_destroy_block(result: *mut types::Block) {
 }
 
 
-
-
-
 /// Register all the callbacks for use across FFI
 
 #[no_mangle]
@@ -89,7 +86,6 @@ pub unsafe extern fn register_processor(
     destroy_hash: HashDestroy,
     destroy_snapshot: LLMQSnapshotDestroy,
     should_process_diff_with_range: ShouldProcessDiffWithRange,
-    send_error: SendError,
     log_message: LogMessage
 ) -> *mut MasternodeProcessor {
     let processor = MasternodeProcessor::new(
@@ -107,7 +103,6 @@ pub unsafe extern fn register_processor(
         destroy_hash,
         destroy_snapshot,
         should_process_diff_with_range,
-        send_error,
         log_message
     );
     println!("register_processor: {:?}", processor);
@@ -167,169 +162,6 @@ pub extern "C" fn process_mnlistdiff_from_message(
     processor.log(format!("process_mnlistdiff_from_message.finish: {:?} {:#?}", std::time::Instant::now(), result));
     boxed(result)
 }
-
-/*
-/// Read message received as a response for 'GETQRINFO' call
-#[no_mangle]
-pub extern "C" fn read_qrinfo(
-    message_arr: *const u8,
-    message_length: usize,
-    processor: *mut MasternodeProcessor,
-    context: *const std::ffi::c_void,
-) -> *mut types::QRInfo {
-    let processor = unsafe { &mut *processor };
-    processor.opaque_context = context;
-    processor.log(format!("read_qrinfo.start: {:?} {:p} {:?}", std::time::Instant::now(), processor, context));
-    let message: &[u8] = unsafe { slice::from_raw_parts(message_arr, message_length as usize) };
-    let block_height_lookup = |hash| processor.lookup_block_height_by_hash(hash);
-    let read_list_diff = |offset: &mut usize| llmq::MNListDiff::new(message, offset, block_height_lookup);
-    let read_snapshot = |offset: &mut usize| types::LLMQSnapshot::from_bytes(message, offset);
-    let read_var_int = |offset: &mut usize| encode::VarInt::from_bytes(message, offset);
-    let offset = &mut 0;
-    let snapshot_at_h_c = boxed(unwrap_or_qr_failure!(read_snapshot(offset)));
-    let snapshot_at_h_2c = boxed(unwrap_or_qr_failure!(read_snapshot(offset)));
-    let snapshot_at_h_3c = boxed(unwrap_or_qr_failure!(read_snapshot(offset)));
-    let mn_list_diff_tip = boxed(unwrap_or_qr_failure!(read_list_diff(offset)).encode());
-    let mn_list_diff_at_h = boxed(unwrap_or_qr_failure!(read_list_diff(offset)).encode());
-    let mn_list_diff_at_h_c = boxed(unwrap_or_qr_failure!(read_list_diff(offset)).encode());
-    let mn_list_diff_at_h_2c = boxed(unwrap_or_qr_failure!(read_list_diff(offset)).encode());
-    let mn_list_diff_at_h_3c = boxed(unwrap_or_qr_failure!(read_list_diff(offset)).encode());
-    let extra_share = message.read_with::<bool>(offset, {}).unwrap_or(false);
-    let (snapshot_at_h_4c, mn_list_diff_at_h_4c) = if extra_share {
-        (boxed(unwrap_or_qr_failure!(read_snapshot(offset))),
-         boxed(unwrap_or_qr_failure!(read_list_diff(offset)).encode()))
-    } else {
-        (null_mut(), null_mut())
-    };
-    let last_quorum_per_index_count = unwrap_or_qr_failure!(read_var_int(offset)).0 as usize;
-    let mut last_quorum_per_index_vec: Vec<*mut types::LLMQEntry> = Vec::with_capacity(last_quorum_per_index_count);
-    for _i in 0..last_quorum_per_index_count {
-        last_quorum_per_index_vec.push(boxed(unwrap_or_qr_failure!(LLMQEntry::from_bytes(message, offset)).encode()));
-    }
-    let quorum_snapshot_list_count = unwrap_or_qr_failure!(read_var_int(offset)).0 as usize;
-    let mut quorum_snapshot_list_vec: Vec<*mut types::LLMQSnapshot> = Vec::with_capacity(quorum_snapshot_list_count);
-    for _i in 0..quorum_snapshot_list_count {
-        quorum_snapshot_list_vec.push(boxed(unwrap_or_qr_failure!(read_snapshot(offset))));
-    }
-    let mn_list_diff_list_count = unwrap_or_qr_failure!(read_var_int(offset)).0 as usize;
-    let mut mn_list_diff_list_vec: Vec<*mut types::MNListDiff> = Vec::with_capacity(mn_list_diff_list_count);
-    for _i in 0..mn_list_diff_list_count {
-        mn_list_diff_list_vec.push(boxed(unwrap_or_qr_failure!(read_list_diff(offset)).encode()));
-    }
-    let result = types::QRInfo {
-        snapshot_at_h_c,
-        snapshot_at_h_2c,
-        snapshot_at_h_3c,
-        snapshot_at_h_4c,
-        mn_list_diff_tip,
-        mn_list_diff_at_h,
-        mn_list_diff_at_h_c,
-        mn_list_diff_at_h_2c,
-        mn_list_diff_at_h_3c,
-        mn_list_diff_at_h_4c,
-        extra_share,
-        last_quorum_per_index_count,
-        last_quorum_per_index: boxed_vec(last_quorum_per_index_vec),
-        quorum_snapshot_list_count,
-        quorum_snapshot_list: boxed_vec(quorum_snapshot_list_vec),
-        mn_list_diff_list_count,
-        mn_list_diff_list: boxed_vec(mn_list_diff_list_vec),
-    };
-    processor.log(format!("read_qrinfo.finish: {:?} {:#?}", std::time::Instant::now(), result));
-    boxed(result)
-}
-
-/// Here we calculate quorums according to Core v0.18
-/// See https://github.com/dashpay/dips/blob/master/dip-0024.md
-#[no_mangle]
-pub extern "C" fn process_qrinfo(
-    info: *mut types::QRInfo,
-    use_insight_as_backup: bool,
-    genesis_hash: *const u8,
-    processor: *mut MasternodeProcessor,
-    cache: *mut MasternodeProcessorCache,
-    context: *const std::ffi::c_void,
-) -> *mut types::QRInfoResult {
-    let llmq_rotation_info = unsafe { *info };
-    let cache = unsafe { &mut *cache };
-    let extra_share = llmq_rotation_info.extra_share;
-    let processor = unsafe { &mut *processor };
-    processor.opaque_context = context;
-    processor.use_insight_as_backup = use_insight_as_backup;
-    processor.genesis_hash = genesis_hash;
-    processor.log(format!("process_qrinfo.start: {:?} {:?} {:p} {:p} {:?}", std::time::Instant::now(), genesis_hash, processor, cache, context));
-    let mut process_list_diff = |list_diff: llmq::MNListDiff| processor.get_list_diff_result_with_base_lookup(list_diff, cache);
-    let decode_diff = |list_diff: *mut types::MNListDiff| unsafe { (*(list_diff)).decode() };
-    let decode_snapshot = |snapshot: *mut types::LLMQSnapshot| unsafe { (*(snapshot)).decode() };
-    let mut get_list_diff_result = |list_diff: llmq::MNListDiff| boxed(process_list_diff(list_diff));
-
-    let diff_at_tip = decode_diff(llmq_rotation_info.mn_list_diff_tip);
-    let diff_at_h = decode_diff(llmq_rotation_info.mn_list_diff_at_h);
-    let diff_at_h_c = decode_diff(llmq_rotation_info.mn_list_diff_at_h_c);
-    let diff_at_h_2c = decode_diff(llmq_rotation_info.mn_list_diff_at_h_2c);
-    let diff_at_h_3c = decode_diff(llmq_rotation_info.mn_list_diff_at_h_3c);
-    let diff_at_h_4c = if extra_share { Some(decode_diff(llmq_rotation_info.mn_list_diff_at_h_4c)) } else { None };
-
-    let snapshot_at_h_c = decode_snapshot(llmq_rotation_info.snapshot_at_h_c);
-    let snapshot_at_h_2c = decode_snapshot(llmq_rotation_info.snapshot_at_h_2c);
-    let snapshot_at_h_3c = decode_snapshot(llmq_rotation_info.snapshot_at_h_3c);
-    let snapshot_at_h_4c = if extra_share { Some(decode_snapshot(llmq_rotation_info.snapshot_at_h_4c)) } else { None };
-
-    processor.save_snapshot(diff_at_h_c.block_hash, snapshot_at_h_c.clone());
-    processor.save_snapshot(diff_at_h_2c.block_hash, snapshot_at_h_2c.clone());
-    processor.save_snapshot(diff_at_h_3c.block_hash, snapshot_at_h_3c.clone());
-
-    if extra_share {
-        processor.save_snapshot(diff_at_h_4c.as_ref().unwrap().block_hash, snapshot_at_h_4c.as_ref().unwrap().clone());
-    }
-
-    let result_at_tip = get_list_diff_result(diff_at_tip);
-    let result_at_h = get_list_diff_result(diff_at_h);
-    let result_at_h_c = get_list_diff_result(diff_at_h_c);
-    let result_at_h_2c = get_list_diff_result(diff_at_h_2c);
-    let result_at_h_3c = get_list_diff_result(diff_at_h_3c);
-    let result_at_h_4c = if extra_share { get_list_diff_result(diff_at_h_4c.unwrap()) } else { null_mut() };
-
-    let last_quorum_per_index_count = llmq_rotation_info.last_quorum_per_index_count;
-    let quorum_snapshot_list_count = llmq_rotation_info.quorum_snapshot_list_count;
-    let mn_list_diff_list_count = llmq_rotation_info.mn_list_diff_list_count;
-    let last_quorum_per_index = llmq_rotation_info.last_quorum_per_index;
-
-    let mut diffs = Vec::<*mut types::MNListDiffResult>::with_capacity(mn_list_diff_list_count);
-    for i in 0..mn_list_diff_list_count {
-        let list_diff_encoded = unsafe { *llmq_rotation_info.mn_list_diff_list.offset(i as isize) };
-        let list_diff = decode_diff(list_diff_encoded);
-        let list_diff_block_hash = list_diff.block_hash;
-        let list_diff_result = get_list_diff_result(list_diff);
-        diffs.push(list_diff_result);
-        let snapshot_encoded = unsafe { *llmq_rotation_info.quorum_snapshot_list.offset(i as isize) };
-        let snapshot_encoded = decode_snapshot(snapshot_encoded);
-        processor.save_snapshot(list_diff_block_hash, snapshot_encoded);
-    }
-
-    let result = types::QRInfoResult {
-        result_at_tip,
-        result_at_h,
-        result_at_h_c,
-        result_at_h_2c,
-        result_at_h_3c,
-        result_at_h_4c,
-        snapshot_at_h_c: llmq_rotation_info.snapshot_at_h_c,
-        snapshot_at_h_2c: llmq_rotation_info.snapshot_at_h_2c,
-        snapshot_at_h_3c: llmq_rotation_info.snapshot_at_h_3c,
-        snapshot_at_h_4c: llmq_rotation_info.snapshot_at_h_4c,
-        extra_share,
-        last_quorum_per_index_count,
-        last_quorum_per_index,
-        quorum_snapshot_list_count,
-        quorum_snapshot_list: llmq_rotation_info.quorum_snapshot_list,
-        mn_list_diff_list_count,
-        mn_list_diff_list: boxed_vec(diffs),
-    };
-    processor.log(format!("process_qrinfo.finish: {:?} {:#?}", std::time::Instant::now(), result));
-    boxed(result)
-}
-*/
 
 /// Here we read & calculate quorums according to Core v0.18
 /// See https://github.com/dashpay/dips/blob/master/dip-0024.md
@@ -442,121 +274,6 @@ pub extern "C" fn process_qrinfo_from_message(
     boxed(result)
 }
 
-/// This is convenience Core v0.17 method for use in tests which doesn't involve cross-FFI calls
-pub fn process_mnlistdiff_from_message_internal(
-    message_arr: *const u8,
-    message_length: usize,
-    use_insight_as_backup: bool,
-    genesis_hash: *const u8,
-    processor: *mut MasternodeProcessor,
-    cache: *mut MasternodeProcessorCache,
-    context: *const std::ffi::c_void,
-) -> MNListDiffResult {
-    let processor = unsafe { &mut *processor };
-    let cache = unsafe { &mut *cache };
-    println!("process_mnlistdiff_from_message_internal.start: {:?}", std::time::Instant::now());
-    processor.opaque_context = context;
-    processor.use_insight_as_backup = use_insight_as_backup;
-    processor.genesis_hash = genesis_hash;
-    let message: &[u8] = unsafe { slice::from_raw_parts(message_arr, message_length as usize) };
-    let list_diff = unwrap_or_diff_processing_failure!(llmq::MNListDiff::new(message, &mut 0, |hash| processor.lookup_block_height_by_hash(hash)));
-    let result = processor.get_list_diff_result_internal_with_base_lookup(list_diff, cache);
-    println!("process_mnlistdiff_from_message_internal.finish: {:?} {:#?}", std::time::Instant::now(), result);
-    result
-
-
-}
-
-/// This is convenience Core v0.18 method for use in tests which doesn't involve cross-FFI calls
-pub fn process_qrinfo_from_message_internal(
-    message: *const u8,
-    message_length: usize,
-    use_insight_as_backup: bool,
-    genesis_hash: *const u8,
-    processor: *mut MasternodeProcessor,
-    cache: *mut MasternodeProcessorCache,
-    context: *const std::ffi::c_void,
-) -> QRInfoResult {
-    println!("process_qrinfo_from_message: {:?} {:?}", processor, cache);
-    let message: &[u8] = unsafe { slice::from_raw_parts(message, message_length as usize) };
-    let processor = unsafe { &mut *processor };
-    processor.opaque_context = context;
-    processor.use_insight_as_backup = use_insight_as_backup;
-    processor.genesis_hash = genesis_hash;
-    let cache = unsafe { &mut *cache };
-    println!("process_qrinfo_from_message --: {:?} {:?} {:?}", processor, processor.opaque_context, cache);
-    let offset = &mut 0;
-    let read_list_diff = |offset: &mut usize|
-        processor.read_list_diff_from_message(message, offset);
-    let mut process_list_diff = |list_diff: llmq::MNListDiff|
-        processor.get_list_diff_result_internal_with_base_lookup(list_diff, cache);
-    let read_snapshot = |offset: &mut usize| llmq::LLMQSnapshot::from_bytes(message, offset);
-    let read_var_int = |offset: &mut usize| encode::VarInt::from_bytes(message, offset);
-    let snapshot_at_h_c = unwrap_or_qr_processing_failure!(read_snapshot(offset));
-    let snapshot_at_h_2c = unwrap_or_qr_processing_failure!(read_snapshot(offset));
-    let snapshot_at_h_3c = unwrap_or_qr_processing_failure!(read_snapshot(offset));
-    let diff_tip = unwrap_or_qr_processing_failure!(read_list_diff(offset));
-    let diff_h = unwrap_or_qr_processing_failure!(read_list_diff(offset));
-    let diff_h_c = unwrap_or_qr_processing_failure!(read_list_diff(offset));
-    let diff_h_2c = unwrap_or_qr_processing_failure!(read_list_diff(offset));
-    let diff_h_3c = unwrap_or_qr_processing_failure!(read_list_diff(offset));
-    let extra_share = message.read_with::<bool>(offset, {}).unwrap_or(false);
-    let (snapshot_at_h_4c, diff_h_4c) = if extra_share {
-        (Some(unwrap_or_qr_processing_failure!(read_snapshot(offset))),
-         Some(unwrap_or_qr_processing_failure!(read_list_diff(offset))))
-    } else {
-        (None, None)
-    };
-    processor.save_snapshot(diff_h_c.block_hash, snapshot_at_h_c.clone());
-    processor.save_snapshot(diff_h_2c.block_hash, snapshot_at_h_2c.clone());
-    processor.save_snapshot(diff_h_3c.block_hash, snapshot_at_h_3c.clone());
-    if extra_share {
-        processor.save_snapshot(diff_h_4c.as_ref().unwrap().block_hash, snapshot_at_h_4c.as_ref().unwrap().clone());
-    }
-    let last_quorum_per_index_count = unwrap_or_qr_processing_failure!(read_var_int(offset)).0 as usize;
-    let mut last_quorum_per_index: Vec<LLMQEntry> = Vec::with_capacity(last_quorum_per_index_count);
-    for _i in 0..last_quorum_per_index_count {
-        last_quorum_per_index.push(unwrap_or_qr_processing_failure!(LLMQEntry::from_bytes(message, offset)));
-    }
-    let quorum_snapshot_list_count = unwrap_or_qr_processing_failure!(read_var_int(offset)).0 as usize;
-    let mut quorum_snapshot_list: Vec<llmq::LLMQSnapshot> = Vec::with_capacity(quorum_snapshot_list_count);
-    for _i in 0..quorum_snapshot_list_count {
-        quorum_snapshot_list.push(unwrap_or_qr_processing_failure!(read_snapshot(offset)));
-    }
-    let mn_list_diff_list_count = unwrap_or_qr_processing_failure!(read_var_int(offset)).0 as usize;
-    let mut mn_list_diff_list: Vec<MNListDiffResult> = Vec::with_capacity(mn_list_diff_list_count);
-    for _i in 0..mn_list_diff_list_count {
-        mn_list_diff_list.push(process_list_diff(unwrap_or_qr_processing_failure!(read_list_diff(offset))));
-    }
-    // The order is important since the each new one dependent on previous
-    let result_at_h_4c = if let Some(diff) = diff_h_4c {
-        Some(process_list_diff(diff))
-    } else {
-        None
-    };
-    let result_at_h_3c = process_list_diff(diff_h_3c);
-    let result_at_h_2c = process_list_diff(diff_h_2c);
-    let result_at_h_c = process_list_diff(diff_h_c);
-    let result_at_h = process_list_diff(diff_h);
-    let result_at_tip = process_list_diff(diff_tip);
-    QRInfoResult {
-        error_status: ProcessingError::None,
-        result_at_tip,
-        result_at_h,
-        result_at_h_c,
-        result_at_h_2c,
-        result_at_h_3c,
-        result_at_h_4c,
-        snapshot_at_h_c,
-        snapshot_at_h_2c,
-        snapshot_at_h_3c,
-        snapshot_at_h_4c,
-        extra_share,
-        last_quorum_per_index,
-        quorum_snapshot_list,
-        mn_list_diff_list
-    }
-}
 
 // #[no_mangle]
 // pub extern "C" fn test_func(get_masternode_list_by_block_hash: MasternodeListLookup, destroy_masternode_list: MasternodeListDestroy, opaque_context: *const std::ffi::c_void) {
