+++ src/tests/llmq_rotation.rs
@@ -1,8 +1,8 @@
 use dash_spv_models::common::chain_type::ChainType;
 use dash_spv_primitives::crypto::byte_util::{Reversable, UInt256};
 use dash_spv_primitives::hashes::hex::{FromHex, ToHex};
-use crate::lib_tests::tests::{add_insight_lookup_default, block_height_lookup_5078, FFIContext, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, masternode_list_destroy_default, get_masternode_list_by_block_hash_default, masternode_list_save_default, message_from_file, save_llmq_snapshot_default, should_process_llmq_of_type, validate_llmq_callback, get_merkle_root_by_hash_default, log_default, get_masternode_list_by_block_hash_from_cache, masternode_list_save_in_cache, save_llmq_snapshot_in_cache, hash_destroy_default, snapshot_destroy_default, should_process_diff_with_range_default, send_error_default};
-use crate::{process_mnlistdiff_from_message_internal, process_qrinfo_from_message, process_qrinfo_from_message_internal, processor_create_cache, register_processor};
+use crate::lib_tests::tests::{add_insight_lookup_default, block_height_lookup_5078, FFIContext, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, masternode_list_destroy_default, get_masternode_list_by_block_hash_default, masternode_list_save_default, message_from_file, save_llmq_snapshot_default, should_process_llmq_of_type, validate_llmq_callback, get_merkle_root_by_hash_default, log_default, get_masternode_list_by_block_hash_from_cache, masternode_list_save_in_cache, save_llmq_snapshot_in_cache, hash_destroy_default, snapshot_destroy_default, should_process_diff_with_range_default, process_qrinfo_from_message_internal, process_mnlistdiff_from_message_internal};
+use crate::{process_qrinfo_from_message, processor_create_cache, register_processor};
 use crate::processing::MasternodeProcessorCache;
 
 // #[test]
@@ -31,7 +31,6 @@ fn test_llmq_rotation() {
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
@@ -78,7 +77,6 @@ fn test_llmq_rotation_2() {
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
@@ -398,7 +396,6 @@ fn test_devnet_333() {
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
@@ -432,7 +429,6 @@ fn test_processor_devnet_333() {
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
@@ -651,7 +647,6 @@ fn test_processor_devnet_333_2() {
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
@@ -932,7 +927,6 @@ fn test_jack_daniels() {
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
