+++ src/tests/llmq_verification.rs
@@ -9,7 +9,7 @@ use dash_spv_models::masternode::LLMQEntry;
 use dash_spv_primitives::crypto::byte_util::{Reversable, UInt256};
 use dash_spv_primitives::hashes::hex::ToHex;
 use crate::{process_mnlistdiff_from_message, processor_create_cache, register_processor};
-use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, block_height_for, block_height_lookup_default, FFIContext, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_from_cache, get_merkle_root_by_hash_default, hash_destroy_default, log_default, masternode_list_destroy_default, masternode_list_save_in_cache, message_from_file, save_llmq_snapshot_default, send_error_default, should_process_diff_with_range_default, should_process_llmq_of_type, snapshot_destroy_default, validate_llmq_callback};
+use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, block_height_for, block_height_lookup_default, FFIContext, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_from_cache, get_merkle_root_by_hash_default, hash_destroy_default, log_default, masternode_list_destroy_default, masternode_list_save_in_cache, message_from_file, save_llmq_snapshot_default, should_process_diff_with_range_default, should_process_llmq_of_type, snapshot_destroy_default, validate_llmq_callback};
 use crate::processing::MasternodeProcessorCache;
 
 
@@ -37,7 +37,6 @@ fn testnet_llmq_verification() { //testTestnetQuorumVerification
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
@@ -134,7 +133,6 @@ fn testnet_llmq_verification_using_processor_and_cache() { //testTestnetQuorumVe
         hash_destroy_default,
         snapshot_destroy_default,
         should_process_diff_with_range_default,
-        send_error_default,
         log_default)
     };
     let cache = unsafe { processor_create_cache() };
