+++ src/tests/mainnet_reload.rs
@@ -3,7 +3,7 @@ use dash_spv_ffi::ffi::from::FromFFI;
 use dash_spv_models::common::chain_type::ChainType;
 use dash_spv_models::masternode;
 use dash_spv_primitives::crypto::byte_util::UInt256;
-use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, block_height_lookup_default, FFIContext, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_from_cache, get_merkle_root_by_hash_default, hash_destroy_default, log_default, masternode_list_destroy_default, masternode_list_save_in_cache, message_from_file, save_llmq_snapshot_default, send_error_default, should_process_diff_with_range_default, should_process_llmq_of_type, snapshot_destroy_default, validate_llmq_callback};
+use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, block_height_lookup_default, FFIContext, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_from_cache, get_merkle_root_by_hash_default, hash_destroy_default, log_default, masternode_list_destroy_default, masternode_list_save_in_cache, message_from_file, save_llmq_snapshot_default, should_process_diff_with_range_default, should_process_llmq_of_type, snapshot_destroy_default, validate_llmq_callback};
 use crate::{process_mnlistdiff_from_message, processor_create_cache, register_processor};
 
 #[test]
@@ -63,7 +63,6 @@ pub fn load_masternode_lists_for_files(files: Vec<String>, chain: ChainType, ass
             hash_destroy_default,
             snapshot_destroy_default,
             should_process_diff_with_range_default,
-            send_error_default,
             log_default,
         )
     };
